added additional tests
